some store-level events work on KeyValueStore
also modified KeyValueStore so that we don't assume arrays are
symmetrical. You can now have a parentRelationship and a
childRelationship that are not inverses of each other in the store,
without breaking reads and writes.

This feature exists because many KeyValueStore objects are caches, so
there can easily be a time where we hot-update one entry and its hasMany
children, but don't hot-update all the inverse children. They will be
eventually consistent (when all the reads resolve).

More testing is needed to make sure this doesn't break expectations when
the KeyValueStore is a terminal (say you're just sitting on a redis),
because for terminal stores, the value has to be canonical. But I'm
pretty sure this happens already - the only way the sides of a hasMany
can be asymmetric is when one half fires an update event (cacheable read
/ write) when the other doesn't for some reason.