streamget in subscribe
I'm not sure if this is the right decision. It *works* for now, and
meets the spec I have in the tests, but it feels a bit hacky.

To wit: should we be doing a streamGet inside the subscribe call?
Should I instead just use streamGet in general as the subscription?
Probably, yes. More work here needed, but for now it's fine.