properly supports invalidation events.
typical use: when the back end notices that there's been a change
to a field (child or self), it emits an event to the front end via
websocket or some other side-channel. pass this information to Plump
and then the data will be deleted from hot caches, and if there are
connected subscribers, a re-read from the terminal store will be
initiated. If there's no connected subscribers, the re-read will happen
on next load.